split the old "characters.txt" into multiple files.
We simply exclude many conversions because the content
of "traditional_characters.txt" also contains some simplified
characters.

This change builds 3 conversions table follows the descriptions
of Unihan_Variants.txt from:

    http://www.unicode.org/reports/tr38/#SCTC

We also produce a full list of characters that has no variants, which
means they are just use as-is in both traditional and simplified
Chinese. (case 1 under 3.7.1 of the abovementioned URL.)